lesson-18-ai/vanna/vanna-streamlit/requirements.txt
## check https://github.com/vanna-ai/vanna/blob/main/pyproject.toml
## for any missing pkg
streamlit-aggrid
lxml
streamlit_option_menu
jsonlines
notebook
ollama
# vanna
# use wgong/ssadata (install from git source) v0.7.4
streamlit
streamlit_code_editor
python-dotenv
google-generativeai
google-cloud-aiplatform
chromadb
openai
mistralai
anthropic
google-generativeai
boto3
